patch:(fix some bugs)

+++ app/src/main/java/com/rezapour/carstask/data/network/model/CarNetwrokEntity.kt
@@ -54,11 +54,11 @@ data class CarNetwrokEntity constructor(
 
     @SerializedName("latitude")
     @Expose
-    val latitude: Long,
+    val latitude: Double,
 
     @SerializedName("longitude")
     @Expose
-    val longitude: Long,
+    val longitude: Double,
 
     @SerializedName("innerCleanliness")
     @Expose
